Split compiler passes
