+++ CmfRoutingAutoBundle.php
@@ -14,15 +14,17 @@ namespace Symfony\Cmf\Bundle\RoutingAutoBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Cmf\Bundle\RoutingAutoBundle\DependencyInjection\Compiler\AutoRoutePass;
 use Doctrine\Bundle\PHPCRBundle\DependencyInjection\Compiler\DoctrinePhpcrMappingsPass;
+use Symfony\Cmf\Bundle\RoutingAutoBundle\DependencyInjection\Compiler\ServicePass;
+use Symfony\Cmf\Bundle\RoutingAutoBundle\DependencyInjection\Compiler\AdapterPass;
 
 class CmfRoutingAutoBundle extends Bundle
 {
     public function build(ContainerBuilder $container)
     {
         parent::build($container);
-        $container->addCompilerPass(new AutoRoutePass());
+        $container->addCompilerPass(new ServicePass());
+        $container->addCompilerPass(new AdapterPass());
         $this->buildPhpcrCompilerPass($container);
     }
